requirements.txt
pdfminer
scikit-image
Spire.Doc
Spire.Pdf
tqdm
parmap
copydetect==0.5.0
